requirements.txt
black==22.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
flask-swagger-ui==4.11.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pymongo==4.2.0
python-dotenv==0.20.0
tomli==2.0.1
Werkzeug==2.2.2
